+++ client/contracts/truffle/scripts/truffleTools/truffleUpdateTool.js
@@ -31,7 +31,7 @@ const path = require('path');
 const JSON5 = require('json5');
 
 const network = global.process.argv[4];
-const source = fs.readFileSync(path.resolve(__dirname, '../truffle.js'), 'utf8');
+const source = fs.readFileSync(path.resolve(__dirname, '../../truffle.js'), 'utf8');
 const truffleJS = JSON5.parse(source.slice(17, -2));
 const host = truffleJS.networks[network]['host'];
 const port = truffleJS.networks[network]['port'];
@@ -43,7 +43,7 @@ const {
   PublicResolver,
   readABIAndByteCode,
   getParamFromTxEvent,
-} = require('./helpers.js');
+} = require('../helpers/helpers.js');
 
 const MultiSig = artifacts.require('MultiSigWallet.sol');
 const multiSigAddr = '0x0e5d0Fd336650E663C710EF420F85Fb081E21415';
