+++ client/contracts/truffle/scripts/utilities/consoleScript.js
